first changes, data for games and introductory screens
