+++ js/blocks/footer.js
@@ -1,14 +1,6 @@
 import createElement from '../createElement';
-import showScreen from '../showScreen';
-import moduleGreeting from './greeting';
 
-const introStr = String.raw`<div id="main" class="central__content">
-    <div id="intro" class="intro">
-      <h1 class="intro__asterisk">*</h1>
-      <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
-    </div>
-  </div>
-  <footer class="footer">
+const footerStr = String.raw`<footer class="footer">
     <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
     <span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
     <div class="footer__social-links">
@@ -19,11 +11,6 @@ const introStr = String.raw`<div id="main" class="central__content">
     </div>
   </footer>`;
 
-const moduleIntro = createElement(introStr);
-const showGreeting = moduleIntro.querySelector(`.intro__asterisk`);
+const moduleFooter = createElement(footerStr);
 
-showGreeting.addEventListener(`click`, () => {
-  showScreen(moduleGreeting);
-});
-
-export default moduleIntro;
+export default moduleFooter;
