+++ js/blocks/games/game-1.js
@@ -1,27 +1,15 @@
-import createElement from '../createElement';
-import showScreen from '../showScreen';
+import createElement from '../../createElement';
+import showScreen from '../../showScreen';
 import moduleSecondGame from './game-2';
-import moduleGreeting from './greeting';
+import moduleGreeting from '../greeting/greeting';
+import headerStr from './header';
+import data from './gameData';
 
-const firstGameStr = String.raw`<header class="header">
-    <div class="header__back">
-      <button class="back">
-        <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-        <img src="img/logo_small.svg" width="101" height="44">
-      </button>
-    </div>
-    <h1 class="game__timer">NN</h1>
-    <div class="game__lives">
-      <img src="img/heart__empty.svg" class="game__heart" alt="Life" width="32" height="32">
-      <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
-      <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
-    </div>
-  </header>
-  <div class="game">
-    <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
+const firstGameStr = String.raw`<div class="game">
+    <p class="game__task">${data.game1.text}</p>
     <form class="game__content">
       <div class="game__option">
-        <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
+        <img src="${data.game1.answers[0].imgSrc}" alt="Option 1" width="468" height="458">
         <label class="game__answer game__answer--photo">
           <input name="question1" type="radio" value="photo">
           <span>Фото</span>
@@ -32,7 +20,7 @@ const firstGameStr = String.raw`<header class="header">
         </label>
       </div>
       <div class="game__option">
-        <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
+        <img src="${data.game1.answers[1].imgSrc}" alt="Option 2" width="468" height="458">
         <label class="game__answer  game__answer--photo">
           <input name="question2" type="radio" value="photo">
           <span>Фото</span>
@@ -57,19 +45,9 @@ const firstGameStr = String.raw`<header class="header">
         <li class="stats__result stats__result--unknown"></li>
       </ul>
     </div>
-  </div>
-  <footer class="footer">
-    <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-    <span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-    <div class="footer__social-links">
-      <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-      <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-      <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-      <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-    </div>
-  </footer>`;
+  </div>`;
 
-const moduleFirstGame = createElement(firstGameStr);
+const moduleFirstGame = createElement(headerStr + firstGameStr);
 const goBack = moduleFirstGame.querySelector(`.back`);
 const form = moduleFirstGame.querySelector(`.game__content`);
 const question1 = moduleFirstGame.querySelectorAll(`input[name=question1]`);
