+++ js/blocks/games/gameData.js
@@ -1,0 +1,43 @@
+export default {
+  game1: {
+    text: `Угадайте для каждого изображения фото или рисунок?`,
+    answers: [
+      {
+        imgSrc: `http://placehold.it/468x458`,
+        imgType: `paint`
+      },
+      {
+        imgSrc: `http://placehold.it/468x458`,
+        imgType: `photo`
+      }
+    ]
+  },
+
+  game2: {
+    text: `Угадай, фото или рисунок?`,
+    answers: [
+      {
+        imgSrc: `http://placehold.it/705x455`,
+        imgType: `paint`
+      }
+    ]
+  },
+
+  game3: {
+    text: `Найдите рисунок среди изображений`,
+    answers: [
+      {
+        imgSrc: `http://placehold.it/304x455`,
+        imgType: `paint`
+      },
+      {
+        imgSrc: `http://placehold.it/304x455`,
+        imgType: `photo`
+      },
+      {
+        imgSrc: `http://placehold.it/304x455`,
+        imgType: `photo`
+      }
+    ]
+  }
+};
