+++ js/blocks/greeting/greetingData.js
@@ -1,0 +1,8 @@
+export default {
+  title: `Лучшие художники-фотореалисты бросают&nbsp;тебе&nbsp;вызов!`,
+  text: `Правила игры просты.<br>
+        Нужно отличить рисунок&nbsp;от фотографии и сделать выбор.<br>
+        Задача кажется тривиальной, но не думай, что все так просто.<br>
+        Фотореализм обманчив и коварен.<br>
+        Помни, главное — смотреть очень внимательно.`
+};
