+++ js/blocks/rules/rulesData.js
@@ -1,0 +1,11 @@
+export default {
+  title: `Правила`,
+  text: `Угадай 10 раз для каждого изображения фото <img
+      src="img/photo_icon.png" width="16" height="16"> или рисунок <img
+      src="img/paint_icon.png" width="16" height="16" alt="">.<br>
+      Фотографиями или рисунками могут быть оба изображения.<br>
+      На каждую попытку отводится 30 секунд.<br>
+      Ошибиться можно не более 3 раз.<br>
+      <br>
+      Готовы?`
+};
