+++ js/blocks/stats/stats.js
@@ -1,6 +1,6 @@
-import createElement from '../createElement';
-import showScreen from '../showScreen';
-import moduleGreeting from './greeting';
+import createElement from '../../createElement';
+import showScreen from '../../showScreen';
+import moduleGreeting from '../greeting/greeting';
 
 const statsStr = String.raw`<header class="header">
     <div class="header__back">
@@ -109,17 +109,7 @@ const statsStr = String.raw`<header class="header">
         <td colspan="5" class="result__total  result__total--final">950</td>
       </tr>
     </table>
-  </div>
-  <footer class="footer">
-    <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-    <span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-    <div class="footer__social-links">
-      <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-      <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-      <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-      <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-    </div>
-  </footer>`;
+  </div>`;
 
 const moduleStats = createElement(statsStr);
 const goBack = moduleStats.querySelector(`.back`);
