+++ js/main.js
@@ -1,4 +1,4 @@
 import showScreen from './showScreen';
-import moduleIntro from './blocks/intro';
+import moduleIntro from './blocks/intro/intro';
 
 showScreen(moduleIntro);
